requirements.txt
Authlib==1.3.2
bidict==0.22.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.0
flask-oidc==2.2.0
Flask-SocketIO==5.3.3
greenlet==3.1.0
gunicorn==23.0.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pytest==8.3.2
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.4
Flask-PyMongo==2.3.0
pydantic==2.9.2
python-dotenv==1.0.1